Separate web and react-native options
