+++ packages/authgear-react-native/src/index.ts
@@ -7,11 +7,8 @@ import {
   _MemoryStorageDriver,
   _StorageDriver,
   _BaseContainer,
-  AuthorizeOptions,
   AuthorizeResult,
-  ReauthenticateOptions,
   ReauthenticateResult,
-  PromoteOptions,
   UserInfo,
   SettingOptions,
   OAuthError,
@@ -36,7 +33,13 @@ import {
   storageSetItem,
   storageDeleteItem,
 } from "./nativemodule";
-import { BiometricOptions, ReactNativeContainerDelegate } from "./types";
+import {
+  BiometricOptions,
+  ReactNativeContainerDelegate,
+  AuthorizeOptions,
+  ReauthenticateOptions,
+  PromoteOptions,
+} from "./types";
 import { getAnonymousJWK, signAnonymousJWT } from "./jwt";
 import { BiometricPrivateKeyNotFoundError } from "./error";
 import { Platform } from "react-native";
