+++ packages/authgear-react-native/src/types.ts
@@ -1,4 +1,4 @@
-import { SessionStateChangeReason } from "@authgear/core";
+import { SessionStateChangeReason, PromptOption } from "@authgear/core";
 import { ReactNativeContainer } from ".";
 
 /**
@@ -30,6 +30,119 @@ export interface ReactNativeContainerDelegate {
    * @public
    */
   sendWechatAuthRequest(state: string): void;
+}
+
+/**
+ * Auth UI authorization options
+ *
+ * @public
+ */
+export interface AuthorizeOptions {
+  /**
+   * Redirect uri. Redirection URI to which the response will be sent after authorization.
+   */
+  redirectURI: string;
+  /**
+   * OAuth 2.0 state value.
+   */
+  state?: string;
+  /**
+   * OIDC prompt parameter.
+   *
+   * Prompt parameter will be used for Authgear authorization, it will also be forwarded to the underlying SSO providers.
+   *
+   * For Authgear, currently, only login is supported. Other unsupported values will be ignored.
+   *
+   * For the underlying SSO providers, some providers only support a single value rather than a list.
+   * The first supported value will be used for that case.
+   * e.g. Azure Active Directory
+   *
+   */
+  prompt?: PromptOption[] | PromptOption;
+  /**
+   * OIDC login hint parameter
+   */
+  loginHint?: string;
+  /**
+   * UI locale tags
+   */
+  uiLocales?: string[];
+  /**
+   * OAuth response type
+   */
+  responseType?: "code" | "none";
+  /**
+   * WeChat Redirect URI is needed when integrating WeChat login in react-native
+   * The wechatRedirectURI will be called when user click the login with WeChat button
+   */
+  wechatRedirectURI?: string;
+  /**
+   * @internal
+   * Platform is provided by the sdk
+   */
+  platform?: string;
+  /**
+   * Initial page to open. Valid values are 'login' and 'signup'.
+   */
+  page?: string;
+}
+
+/**
+ * Options for reauthentication
+ * @public
+ */
+export interface ReauthenticateOptions {
+  /**
+   * Redirect uri. Redirection URI to which the response will be sent after authorization.
+   */
+  redirectURI: string;
+  /**
+   * OAuth 2.0 state value.
+   */
+  state?: string;
+  /**
+   * UI locale tags
+   */
+  uiLocales?: string[];
+  /**
+   * WeChat Redirect URI is needed when integrating WeChat login in react-native
+   * The wechatRedirectURI will be called when user click the login with WeChat button
+   */
+  wechatRedirectURI?: string;
+  /**
+   * @internal
+   * Platform is provided by the sdk
+   */
+  platform?: string;
+  /**
+   * OIDC max_age
+   */
+  maxAge?: number;
+}
+
+/**
+ * Auth UI anonymous user promotion options
+ *
+ * @public
+ */
+export interface PromoteOptions {
+  /**
+   * Redirect uri. Redirection URI to which the response will be sent after authorization.
+   */
+  redirectURI: string;
+  /**
+   * OAuth 2.0 state value.
+   */
+  state?: string;
+  /**
+   * UI locale tags
+   */
+  uiLocales?: string[];
+  /**
+   * WeChat Redirect URI is needed when integrating WeChat login in react-native
+   * The wechatRedirectURI will be called when user click the login with WeChat button
+   */
+  wechatRedirectURI?: string;
 }
 
 /**
